CLI, timer, background refresh, real-time file read
We now have the beginnings of a useful badge. The frame rate
is now determined by a timer. The screen refresh now happens
without a busy-wait so we can get on with drawing the next
frame. Double-buffering is now in play.

The display data now comes directly from a file, reading only
the bits that are required for each frame. A complete frame is
read, but in pieces. The image must be exactly 64 pixels high
but can be any width as long as it is at least 128 pixels.
The width is now determined from the file size.

A rudimentary CLI now exists which allows listing the files
on the badge, removing files and seeing how much space is
in use, as well as starting a zmodem session. A locking
mechanism protects the filesystem from concurrent access.
The display update stops before any command is executed,
or any zmodem activity occurs.

A CLI active screen shows whenever there is at least one
character in the command buffer.

An error screen shows if it's not possible to display data from
a file.

Autobaud now only requires the start bit and LSB.
Unfortunately this does mean that the first character received
on the UART has to have the LSB set for autobaud to work.