Deal with multiple files and workaround for buggy displays.
Firstly, add the ability to choose which file to display based
on its position in the directory. Start at the beginning at
power-up, and allow the right and left buttons to act as next
and previous file selectors, with wraparound. Note that the
buttons aren't properly debounced yet, so that's a thing for
later.

Then, use the up and down buttons to set the display contrast.

While we're at it, set the start position in display memory
at the end of every frame so that if the I2C bus is temporarily
disrupted (such as by shorting against an item of clothing)
then it will recover its position at the next frame.

... and then, implement a workaround for some buggy displays.
Turns out that the latest batch, as well as having slightly
different physical dimensions and positions of the holes,
these displays also wrap around after only 7/8th of the
screen memory has been written. The workaround is to start
writing 7/8ths of the way through so that it wraps at the
real end, and to set the display start line such that it is
still displayed correctly on the screen.