Move UART receive and timeout to interrupt
Also, let's have a more graphical indication that a file transfer
has paused to reassure the user that the badge is on the case.

And just in case we end up with data so corrupt that it crashes the
badge (which shouldn't happen, but who knows...) then we can have
a safe start function such that the badge will not load any media
if the button is pressed at startup.

Because if it saves even one person, it's worth it.