Proper system timer and button debouncing
Now we can use the timer to implement a real delay function
rather than the cycle counting we've been using so far.

And while we're at it, change the speaker frequencies to be
in units of Hz, taking note that the frequency so far is the
frequency of toggling, not the cycle frequency.