Menu system, settings system, fix debouncing
Introducing a menu system, activated by pressing the centre button.

The menu allows viewing the files on the device, and selecting
or deleting them.

A short information screen is given to prompt the user to realise
that they can connect with the serial port and do things.

A settings system is introduced, although currently it doesn't
actually do much with the settings, except for screen brightness.
Settings are not yet written to the filesystem, but hooks are in
place to read and write so that these can be implemented later.

A hook is in place to write a name editor, although the editor
itself is not yet implemented.