Battery meter, save settings, set default file, beep options
Also some refactoring of the way we select files along the way.

Now you can set a default file. This won't be stable across adding
or removing files, but you can always set the default file again.

The beep setting now does something. The battery meter setting now
does something too.

Saving the settings now does something, and the settings are loaded
at startup. The .settings file is still shown in the file menu so
that it can be deleted.